[BE] 상품 수정 리팩토링 (#64)
* refactor: 상품 수정 데이터가 없을 경우 null로 데이터를 처리하도록 코드 리팩토링 #63

* refactor: 상품 수정 데이터가 없을 경우 null로 데이터를 처리하도록 코드 리팩토링 #63

+++ be/src/main/java/team05a/secondhand/product/service/ProductService.java
@@ -83,7 +83,7 @@ public class ProductService {
 	public ProductIdResponse update(ProductUpdateRequest productUpdateRequest, Long productId, Long memberId) {
 		validateProductSeller(productId, memberId);
 		imageService.deleteAllBy(productUpdateRequest.getDeletedImageIds());
-		Long imageCount = imageService.countImagesBy(productId);
+		int imageCount = imageService.countImagesBy(productId);
 		List<String> newImageUrls = imageService.uploadNew(imageCount, productUpdateRequest.getNewImages());
 		String thumbnailUrl = getThumbnailUrl(imageCount, newImageUrls, productId);
 		Product updateProduct = updateProduct(productUpdateRequest, productId, thumbnailUrl);
@@ -91,7 +91,7 @@ public class ProductService {
 		return ProductIdResponse.from(updateProduct.getId());
 	}
 
-	private String getThumbnailUrl(Long imageCount, List<String> newImageUrls, Long productId) {
+	private String getThumbnailUrl(int imageCount, List<String> newImageUrls, Long productId) {
 		if (imageCount == 0) {
 			return newImageUrls.get(0);
 		}
